[LEMS-4477/default-options] AI review
