+++ packages/perseus-core/src/widgets/image/index.ts
@@ -18,9 +18,10 @@ const defaultWidgetOptions: PerseusImageWidgetOptions = {
     alt: "",
     caption: "",
     longDescription: "",
-    // Images are not decorative by default: an image with no alt text is
-    // treated as inaccessible (see `accessible` below) until an author either
-    // writes alt text or explicitly marks the image as decorative.
+    // Marking an image decorative (rendering it with `alt=""` and no caption,
+    // title, or long description) is an explicit author choice, so it starts
+    // off. Note these defaults are inaccessible either way until a background
+    // image is chosen -- see `accessible` below.
     decorative: false,
 };
