+++ packages/perseus-core/src/widgets/orderer/index.ts
@@ -1,17 +1,12 @@
-import {getOrdererPublicWidgetOptions} from "./orderer-util";
+import {
+    getOrdererPublicWidgetOptions,
+    mergeCards,
+    toCard,
+} from "./orderer-util";
 
 import type {OrdererPublicWidgetOptions} from "./orderer-util";
-import type {
-    PerseusOrdererWidgetOptions,
-    PerseusRenderer,
-} from "../../data-schema";
+import type {PerseusOrdererWidgetOptions} from "../../data-schema";
 import type {WidgetLogic} from "../logic-export.types";
-
-const toCard = (content: string): PerseusRenderer => ({
-    content,
-    widgets: {},
-    images: {},
-});
 
 const defaultCorrectOptions = [toCard("$x$")];
 const defaultOtherOptions = [toCard("$y$")];
@@ -19,9 +14,9 @@ const defaultOtherOptions = [toCard("$y$")];
 const defaultWidgetOptions: PerseusOrdererWidgetOptions = {
     correctOptions: defaultCorrectOptions,
     otherOptions: defaultOtherOptions,
-    // `options` is the card bank the learner picks from, which the editor keeps in
-    // sync as the union of the correct answer and the distractors.
-    options: [...defaultCorrectOptions, ...defaultOtherOptions],
+    // `options` is the deck the student picks from, derived the same way the
+    // editor re-derives it on every edit.
+    options: mergeCards(defaultCorrectOptions, defaultOtherOptions),
     height: "normal",
     layout: "horizontal",
 };
