+++ packages/perseus-editor/src/widgets/orderer-editor/orderer-editor.test.tsx
@@ -9,7 +9,7 @@ import * as React from "react";
 
 import {testDependencies} from "../../testing/test-dependencies";
 
-import OrdererEditor, {mergeCards} from "./orderer-editor";
+import OrdererEditor from "./orderer-editor";
 
 import type {PerseusOrdererWidgetOptions} from "@khanacademy/perseus-core";
 import type {UserEvent} from "@testing-library/user-event";
@@ -214,83 +214,5 @@ describe("OrdererEditor", () => {
 
         // Assert
         expect(onChangeMock).toHaveBeenCalledWith({height: "auto"});
-    });
-});
-
-describe("mergeCards", () => {
-    it("combines the correct cards and the other cards", () => {
-        // Arrange, Act
-        const cards = mergeCards(
-            [generateOrdererOption("a")],
-            [generateOrdererOption("b")],
-        );
-
-        // Assert
-        expect(cards).toEqual([
-            generateOrdererOption("a"),
-            generateOrdererOption("b"),
-        ]);
-    });
-
-    it("sorts content with numbers ahead of content without", () => {
-        // Arrange, Act
-        const cards = mergeCards(
-            [generateOrdererOption("3"), generateOrdererOption("$b$")],
-            [generateOrdererOption("2"), generateOrdererOption("a")],
-        );
-
-        // Assert
-        expect(cards).toEqual([
-            generateOrdererOption("2"),
-            generateOrdererOption("3"),
-            generateOrdererOption("$b$"),
-            generateOrdererOption("a"),
-        ]);
-    });
-
-    it("sorts bare variables last", () => {
-        // Arrange, Act
-        const cards = mergeCards(
-            [generateOrdererOption("x + 1"), generateOrdererOption("$y$")],
-            [generateOrdererOption("hello world")],
-        );
-
-        // Assert
-        expect(cards).toEqual([
-            generateOrdererOption("x + 1"),
-            generateOrdererOption("hello world"),
-            generateOrdererOption("$y$"),
-        ]);
-    });
-
-    it("removes duplicate cards", () => {
-        // Arrange, Act
-        const cards = mergeCards(
-            [
-                generateOrdererOption("duplicate"),
-                generateOrdererOption("unique"),
-            ],
-            [generateOrdererOption("duplicate")],
-        );
-
-        // Assert
-        expect(cards).toEqual([
-            generateOrdererOption("duplicate"),
-            generateOrdererOption("unique"),
-        ]);
-    });
-
-    it("removes empty cards", () => {
-        // Arrange, Act
-        const cards = mergeCards(
-            [generateOrdererOption(""), generateOrdererOption("existing")],
-            [generateOrdererOption(""), generateOrdererOption("valid")],
-        );
-
-        // Assert
-        expect(cards).toEqual([
-            generateOrdererOption("existing"),
-            generateOrdererOption("valid"),
-        ]);
     });
 });
